Visualizations of Bike Ride Durations
This the first batch of Google Data Analytics visualizations for the bike sharing company case study. These visualizations show the duration of bike rides for each of the seven days of the week grouped by each of the two user types: member and casual. Members purchase annual subscriptions while those referred to as "casual" do not.